rename useRoute to useZodiacRoute
