+++ extension/src/chains/const.ts
@@ -2,6 +2,8 @@
 
 import { ChainId } from 'ser-kit'
 
+export const ETH_ZERO_ADDRESS = 'eth:0x0000000000000000000000000000000000000000'
+
 export const RPC: Record<ChainId, string> = {
   1: 'https://airlock.gnosisguild.org/api/v1/1/rpc',
   10: 'https://airlock.gnosisguild.org/api/v1/10/rpc',
