+++ extension/src/panel/app-routes/transactions/RolePermissionCheck.tsx
@@ -2,7 +2,7 @@ import { Flex, Tag } from '@/components'
 import { useProvider } from '@/providers'
 import { TransactionState } from '@/state'
 import { Eip1193Provider, JsonRpcError, Route } from '@/types'
-import { useRoute } from '@/zodiac-routes'
+import { useZodiacRoute } from '@/zodiac-routes'
 import { MetaTransactionData } from '@safe-global/safe-core-sdk-types'
 import { toQuantity, ZeroAddress } from 'ethers'
 import React, { useEffect, useState } from 'react'
@@ -82,7 +82,7 @@ const RolePermissionCheck: React.FC<{
   mini?: boolean
 }> = ({ transactionState, index, mini = false }) => {
   const [error, setError] = useState<string | undefined | false>(undefined)
-  const { route } = useRoute()
+  const { route } = useZodiacRoute()
   const provider = useProvider()
 
   const translationAvailable = !!useApplicableTranslation(
