+++ extension/src/panel/integrations/safe/useSafeDelegates.ts
@@ -1,5 +1,5 @@
 import { validateAddress } from '@/utils'
-import { useRoute } from '@/zodiac-routes'
+import { useZodiacRoute } from '@/zodiac-routes'
 import { useEffect, useState } from 'react'
 import { ChainId } from 'ser-kit'
 import { initSafeApiKit } from './kits'
@@ -8,7 +8,7 @@ export const useSafeDelegates = (
   safeAddress: string,
   connectionId?: string
 ) => {
-  const { provider, chainId } = useRoute(connectionId)
+  const { provider, chainId } = useZodiacRoute(connectionId)
 
   const [loading, setLoading] = useState(false)
   const [delegates, setDelegates] = useState<string[]>([])
