+++ extension/src/panel/providers/ProvideProvider.tsx
@@ -1,5 +1,5 @@
 import { Eip1193Provider } from '@/types'
-import { useRoute } from '@/zodiac-routes'
+import { useZodiacRoute } from '@/zodiac-routes'
 import { MetaTransactionData } from '@safe-global/safe-core-sdk-types'
 import { AbiCoder, BrowserProvider, id, TransactionReceipt } from 'ethers'
 import {
@@ -40,7 +40,7 @@ const SubmitTransactionsContext = createContext<
 export const useSubmitTransactions = () => useContext(SubmitTransactionsContext)
 
 export const ProvideProvider = ({ children }: PropsWithChildren) => {
-  const { provider, route, chainId } = useRoute()
+  const { provider, route, chainId } = useZodiacRoute()
   const dispatch = useDispatch()
   const transactions = useTransactions()
