+++ extension/src/panel/zodiac-routes/index.ts
@@ -1,1 +1,2 @@
 export * from './routeHooks'
+export { useZodiacRoute } from './useZodiacRoute'
