+++ extension/src/panel/zodiac-routes/routeHooks.tsx
@@ -1,13 +1,4 @@
-import { getChainId } from '@/chains'
-import {
-  getEip1193ReadOnlyProvider,
-  isConnected,
-  useInjectedWallet,
-  useWalletConnect,
-} from '@/providers'
-import { Eip1193Provider, ProviderType, Route } from '@/types'
-import { ZeroAddress } from 'ethers'
-import { nanoid } from 'nanoid'
+import { Route } from '@/types'
 import React, {
   createContext,
   ReactNode,
@@ -17,7 +8,6 @@ import React, {
   useMemo,
   useRef,
 } from 'react'
-import { parsePrefixedAddress } from 'ser-kit'
 import { useStorage, useStorageEntries } from '../utils'
 
 type RouteContextT = readonly [
@@ -28,16 +18,6 @@ type RouteContextT = readonly [
 const RouteContext = createContext<RouteContextT | null>(null)
 type SelectedRouteContextT = readonly [string, React.Dispatch<string>]
 const SelectedRouteContext = createContext<SelectedRouteContextT | null>(null)
-
-const ETH_ZERO_ADDRESS = 'eth:0x0000000000000000000000000000000000000000'
-const INITIAL_DEFAULT_ROUTE: Route = {
-  id: nanoid(),
-  label: '',
-  providerType: ProviderType.InjectedWallet,
-  avatar: ETH_ZERO_ADDRESS,
-  initiator: undefined,
-  waypoints: undefined,
-}
 
 export const ProvideRoutes: React.FC<{ children: ReactNode }> = ({
   children,
@@ -111,93 +91,4 @@ export const useSelectedRouteId = () => {
     throw new Error('useSelectedRouteId must be used within a <ProvideRoutes>')
   }
   return result
-}
-
-export const useRoute = (id?: string) => {
-  const [routes] = useRoutes()
-  const [selectedRouteId] = useSelectedRouteId()
-  const routeId = id || selectedRouteId
-  const route =
-    (routeId && routes.find((c) => c.id === routeId)) ||
-    routes[0] ||
-    INITIAL_DEFAULT_ROUTE
-
-  const chainId = getChainId(route.avatar)
-
-  const injectedWallet = useInjectedWallet()
-  const walletConnect = useWalletConnect(route.id)
-  const defaultProvider = getEip1193ReadOnlyProvider(chainId)
-
-  const provider: Eip1193Provider =
-    (route.providerType === ProviderType.InjectedWallet
-      ? injectedWallet.provider
-      : walletConnect?.provider) || defaultProvider
-
-  const connected =
-    route.initiator != null &&
-    isConnected(
-      route.providerType === ProviderType.InjectedWallet
-        ? injectedWallet
-        : walletConnect,
-      route.initiator,
-      chainId
-    )
-
-  const providerChainId =
-    route.providerType === ProviderType.InjectedWallet
-      ? injectedWallet.chainId
-      : walletConnect?.chainId || null
-
-  const mustConnectInjectedWallet =
-    route.providerType === ProviderType.InjectedWallet &&
-    !injectedWallet.chainId &&
-    injectedWallet.connected
-
-  const pilotAddress =
-    route.initiator && route.initiator !== `eoa:` + ZeroAddress
-      ? parsePrefixedAddress(route.initiator)[1].toLowerCase()
-      : undefined
-  const canEstablishConnection =
-    !connected &&
-    pilotAddress &&
-    route.providerType === ProviderType.InjectedWallet &&
-    injectedWallet.accounts.some((acc) => acc.toLowerCase() === pilotAddress) &&
-    injectedWallet.chainId !== chainId
-
-  const connectInjectedWallet = injectedWallet.connect
-  const switchChain = injectedWallet.switchChain
-  const requiredChainId = chainId
-
-  useEffect(() => {
-    if (mustConnectInjectedWallet) {
-      connectInjectedWallet()
-    }
-  }, [mustConnectInjectedWallet, connectInjectedWallet])
-
-  const connect = useCallback(async () => {
-    if (requiredChainId && providerChainId !== requiredChainId) {
-      try {
-        await switchChain(requiredChainId)
-      } catch (e) {
-        console.error('Error switching chain', e)
-        return false
-      }
-    }
-
-    return true
-  }, [switchChain, providerChainId, requiredChainId])
-
-  return {
-    route,
-    provider,
-    /** Indicates if `provider` is connected to the right account and chain. */
-    connected,
-
-    chainId,
-    /** The chain ID the `provider` is currently connected to. */
-    providerChainId,
-
-    /** If this callback is set, it can be invoked to establish a connection to the Pilot wallet by asking the user to switch it to the right chain. */
-    connect: canEstablishConnection ? connect : null,
-  }
 }
